BUGFIX: Provide fallback for async ImageTag
With Neos 8.4 I created a thumbnail to be able to generate the width and height for the image attributes.
But in case of `async = true` the initially created thumbnail is async and has no resource and the ResourceUri throws an error.

So we provide a fallback for the first time the thumbnail is requested in which the image size cannot be calculated. After a reload everything is back to normal and the thumbnail and its size are usable.

Resolves: #5717